requirements.txt
# ContextPilot - Open Source Context Augmentation Layer
# Backend Dependencies

# AI/ML
google-genai>=0.4.0
pinecone>=5.0.0

# Crawling
firecrawl-py>=1.0.0
qdrant-client>=1.8.0
trafilatura>=1.12.0

# MCP & API
fastmcp>=0.1.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Database
google-cloud-firestore>=2.16.0
firebase-admin>=6.5.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.26.0
pyyaml>=6.0.0

# Scheduling
apscheduler>=3.10.0
